Add waiting_duration in seconds to model; refactor lib
refactor module BuergerBuero
add waiting_duration to save the amount in seconds for better analysis
