+++ lib/buerger_buero.rb
@@ -1,0 +1,6 @@
+require_relative 'buerger_buero/scraper'
+require_relative 'buerger_buero/crawler'
+require_relative 'buerger_buero/duration'
+
+module BuergerBuero
+end
