+++ lib/buerger_buero/crawler.rb
@@ -1,24 +1,7 @@
 require 'wombat'
-require 'data_mapper'
-require './lib/waiting_citizen'
+require_relative 'duration'
 
 module BuergerBuero
-  class Scraper
-
-    def initialize
-      DataMapper::setup(:default, "sqlite3:///srv/buerberbuero.db")
-      DataMapper.auto_upgrade!
-    end
-
-    def save
-      data = Crawler.new.crawl
-      if data["aktualisiert"]
-        wartende = ::WaitingCitizen.first_or_create({aktualisiert: data["aktualisiert"]}, data )
-        p wartende.inspect
-      end
-    end
-
-  end
   class Crawler
     include Wombat::Crawler
 
@@ -26,6 +9,9 @@ module BuergerBuero
     path "/stadt/buergeramt/mobil-wartezeit.shtml"
 
     wartezeit 'xpath=//*[@id="seite"]/div[2]/p[2]/strong'
+    waiting_duration 'xpath=//*[@id="seite"]/div[2]/p[2]/strong' do |text|
+      Duration.new(text).to_seconds
+    end
     warteende 'xpath=//*[@id="seite"]/div[2]/p[1]/strong'
     naechste_nummer 'xpath=//*[@id="seite"]/div[2]/p[3]/strong'
     aktualisiert 'xpath=//*[@id="seite"]/div[2]/p[4]' do |text|
